+++ TransportationProblem/Domain/Supplier.cs
@@ -13,11 +13,11 @@
         /// <summary>
         /// Доступный объем поставок поставщика.
         /// </summary>
-        public double Capacity { get; set; }
+        public int Capacity { get; set; }
 
         /// <summary>
         /// Значение высланного объема, которое в итоге должно стать равным доступному объёму.
         /// </summary>
-        public double Sent { get; set; }
+        public int Sent { get; set; }
     }
 }
